Refactor modules structure
