+++ src/wasm/utils.rs
@@ -1,7 +1,6 @@
 #[cfg(feature = "wasm_dev")]
 extern crate web_sys;
 
-#[macro_export]
 macro_rules! console_log {
     ( $( $t:tt )* ) => {
         #[cfg(feature = "wasm_dev")]
@@ -13,3 +12,5 @@ pub fn set_panic_hook() {
     #[cfg(feature = "wasm_dev")]
     console_error_panic_hook::set_once();
 }
+
+pub(crate) use console_log;
